requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
pathlib2>=2.3.7
